Fix various 32/64 issues.
- M-Audio driver asio registration isn't proper in Wow6432Node.
  Added a workaround.
- Some ioctl structs weren't bitness independent, so would fail in wow64
- Fix some warnings that show up on 32bit

+++ SarAsio/configui.cpp
@@ -143,7 +143,7 @@ void EndpointsPropertySheetPage::onHardwareInterfaceChanged()
 {
     auto index = ComboBox_GetCurSel(_hardwareInterfaceDropdown) - 1;
 
-    if (index >= 0 && index < _drivers.size()) {
+    if (index >= 0 && index < (int)_drivers.size()) {
         _config.driverClsid = _drivers[index].clsid;
     } else {
         _config.driverClsid = "";
